requirements.txt
streamlit
beautifulsoup4
selenium
opencv-python
requests
pillow
spacy
scikit-image